+++ frame/contracts/src/tests.rs
@@ -16,10 +16,9 @@
 // limitations under the License.
 
 mod pallet_dummy;
+mod unsafe_debug;
 
 use self::test_utils::{ensure_stored, expected_deposit, hash};
-#[cfg(feature = "unsafe-debug")]
-use crate::unsafe_debug::{ExecutionObserver, ExportedFunction};
 use crate::{
 	self as pallet_contracts,
 	chain_extension::{
@@ -52,8 +51,6 @@ use frame_support::{
 };
 use frame_system::{EventRecord, Phase};
 use pallet_contracts_primitives::CodeUploadReturnValue;
-#[cfg(feature = "unsafe-debug")]
-use pallet_contracts_primitives::ExecReturnValue;
 use pretty_assertions::{assert_eq, assert_ne};
 use sp_core::ByteArray;
 use sp_io::hashing::blake2_256;
@@ -63,8 +60,6 @@ use sp_runtime::{
 	traits::{BlakeTwo256, Convert, Hash, IdentityLookup},
 	AccountId32, BuildStorage, Perbill, TokenError,
 };
-#[cfg(feature = "unsafe-debug")]
-use std::cell::RefCell;
 use std::ops::Deref;
 
 type Block = frame_system::mocking::MockBlock<Test>;
@@ -445,52 +440,6 @@ parameter_types! {
 	pub static UnstableInterface: bool = true;
 }
 
-#[cfg(feature = "unsafe-debug")]
-#[derive(Clone, PartialEq, Eq, Debug)]
-struct DebugFrame {
-	code_hash: CodeHash<Test>,
-	call: ExportedFunction,
-	input: Vec<u8>,
-	result: Option<Vec<u8>>,
-}
-
-#[cfg(feature = "unsafe-debug")]
-thread_local! {
-	static DEBUG_EXECUTION_TRACE: RefCell<Vec<DebugFrame>> = RefCell::new(Vec::new());
-}
-
-#[cfg(feature = "unsafe-debug")]
-pub struct TestDebugger;
-#[cfg(feature = "unsafe-debug")]
-impl ExecutionObserver<CodeHash<Test>> for TestDebugger {
-	fn before_call(code_hash: &CodeHash<Test>, entry_point: ExportedFunction, input_data: &[u8]) {
-		DEBUG_EXECUTION_TRACE.with(|d| {
-			d.borrow_mut().push(DebugFrame {
-				code_hash: code_hash.clone(),
-				call: entry_point,
-				input: input_data.to_vec(),
-				result: None,
-			})
-		});
-	}
-
-	fn after_call(
-		code_hash: &CodeHash<Test>,
-		entry_point: ExportedFunction,
-		input_data: Vec<u8>,
-		output: &ExecReturnValue,
-	) {
-		DEBUG_EXECUTION_TRACE.with(|d| {
-			d.borrow_mut().push(DebugFrame {
-				code_hash: code_hash.clone(),
-				call: entry_point,
-				input: input_data,
-				result: Some(output.data.clone()),
-			})
-		});
-	}
-}
-
 impl Config for Test {
 	type Time = Timestamp;
 	type Randomness = Randomness;
@@ -517,7 +466,7 @@ impl Config for Test {
 	type CodeHashLockupDepositPercent = CodeHashLockupDepositPercent;
 	type MaxDelegateDependencies = MaxDelegateDependencies;
 	#[cfg(feature = "unsafe-debug")]
-	type Debug = TestDebugger;
+	type Debug = unsafe_debug::TestDebugger;
 }
 
 pub const ALICE: AccountId32 = AccountId32::new([1u8; 32]);
@@ -5935,96 +5884,6 @@ fn root_cannot_instantiate() {
 				vec![],
 			),
 			DispatchError::RootNotAllowed
-		);
-	});
-}
-
-#[cfg(feature = "unsafe-debug")]
-#[test]
-fn unsafe_debugging_works() {
-	let (wasm_caller, code_hash_caller) = compile_module::<Test>("call").unwrap();
-	let (wasm_callee, code_hash_callee) = compile_module::<Test>("store_call").unwrap();
-
-	fn current_stack() -> Vec<DebugFrame> {
-		DEBUG_EXECUTION_TRACE.with(|stack| stack.borrow().clone())
-	}
-
-	fn deploy(wasm: Vec<u8>) -> AccountId32 {
-		Contracts::bare_instantiate(
-			ALICE,
-			0,
-			GAS_LIMIT,
-			None,
-			Code::Upload(wasm),
-			vec![],
-			vec![],
-			DebugInfo::Skip,
-			CollectEvents::Skip,
-		)
-		.result
-		.unwrap()
-		.account_id
-	}
-
-	fn constructor_frame(hash: CodeHash<Test>, after: bool) -> DebugFrame {
-		DebugFrame {
-			code_hash: hash,
-			call: ExportedFunction::Constructor,
-			input: vec![],
-			result: if after { Some(vec![]) } else { None },
-		}
-	}
-
-	fn call_frame(hash: CodeHash<Test>, args: Vec<u8>, after: bool) -> DebugFrame {
-		DebugFrame {
-			code_hash: hash,
-			call: ExportedFunction::Call,
-			input: args,
-			result: if after { Some(vec![]) } else { None },
-		}
-	}
-
-	use frame_support::traits::Currency;
-
-	ExtBuilder::default().existential_deposit(200).build().execute_with(|| {
-		let _ = Balances::deposit_creating(&ALICE, 1_000_000);
-
-		assert_eq!(current_stack(), vec![]);
-
-		let addr_caller = deploy(wasm_caller);
-		let addr_callee = deploy(wasm_callee);
-
-		assert_eq!(
-			current_stack(),
-			vec![
-				constructor_frame(code_hash_caller, false),
-				constructor_frame(code_hash_caller, true),
-				constructor_frame(code_hash_callee, false),
-				constructor_frame(code_hash_callee, true),
-			]
-		);
-
-		let main_args = (100u32, &addr_callee).encode();
-		let inner_args = (100u32).encode();
-
-		assert_ok!(Contracts::call(
-			RuntimeOrigin::signed(ALICE),
-			addr_caller,
-			0,
-			GAS_LIMIT,
-			None,
-			main_args.clone()
-		));
-
-		let stack_top = current_stack()[4..].to_vec();
-		assert_eq!(
-			stack_top,
-			vec![
-				call_frame(code_hash_caller, main_args.clone(), false),
-				call_frame(code_hash_callee, inner_args.clone(), false),
-				call_frame(code_hash_callee, inner_args, true),
-				call_frame(code_hash_caller, main_args, true),
-			]
 		);
 	});
 }
